+++ examples/read_write.c
@@ -6,10 +6,10 @@
 #include <stdio.h>
 #include <string.h>
 
-#define Read_IFACE iMethod(size_t, read, void *self, char *dest, size_t bytes_to_read)
+#define Read_IFACE vfunc(size_t, read, void *self, char *dest, size_t bytes_to_read)
 interface(Read);
 
-#define Write_IFACE iMethod(size_t, write, void *self, const char *src, size_t bytes_to_write)
+#define Write_IFACE vfunc(size_t, write, void *self, const char *src, size_t bytes_to_write)
 interface(Write);
 
 typedef struct {