+++ examples/trace_state.c
@@ -5,8 +5,8 @@
 #include <stdio.h>
 
 #define State_IFACE                                                                                \
-    iMethod(int, get, void *self)                                                                  \
-    iMethod(void, set, void *self, int x)
+    vfunc(int, get, void *self)                                                                    \
+    vfunc(void, set, void *self, int x)
 
 interface(State);