+++ CHANGELOG.md
@@ -10,6 +10,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
  - `VCALL`, `VCALL_OBJ`, `VCALL_SUPER`, `VCALL_SUPER_OBJ` as a syntax sugar to call virtual functions.
 
+### Changed
+
+ - `iMethod` => `vfunc`, `defaultIMethod` => `defaultVFunc` (consistent with `VTABLE`, `vptr`).
+
 ## [0.7.0] - 2021-10-06
 
 ### Changed